added alerts

+++ view/admin/admin_crud/index.php
@@ -6,9 +6,13 @@ if (!isset($_SESSION['admin'])) {
 }
 include ('../layouts/header.php');
 include ("../../../config.php");
+
+
 ?>
 
 <div class="container-fluid">
+<?php    include ('../layouts/message.php');?>
+
     <div class="row">
         <form action="index.php" method="GET">
             <div class="d-sm-flex d-block align-items-center justify-content mb-9">
@@ -109,7 +113,7 @@ include ("../../../config.php");
                                                     class="btn btn-outline-warning m-1">Edit</a>
                                             </td>
                                             <td class="border-bottom-0">
-                                                <a href="delete.php?admin_id=<?php echo $row["admin_id"] ?>"
+                                                <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?admin_id=<?php echo $row["admin_id"] ?>"
                                                     class="btn btn-outline-danger m-1">Delete</a>
                                             </td>
 
@@ -174,7 +178,7 @@ include ("../../../config.php");
                                                         class="btn btn-outline-warning m-1">Edit</a>
                                                 </td>
                                                 <td class="border-bottom-0">
-                                                    <a href="delete.php?admin_id=<?php echo $row["admin_id"] ?>"
+                                                     <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?admin_id=<?php echo $row["admin_id"] ?>"
                                                         class="btn btn-outline-danger m-1">Delete</a>
                                                 </td>
