+++ view/admin/formation_category_crud/index.php
@@ -81,7 +81,7 @@ include ("../../../config.php");
                                                     class="btn btn-outline-warning m-1">Edit</a>
                                             </td>
                                             <td class="border-bottom-0">
-                                                <a href="delete.php?formation_category_id=<?php echo $row["formation_category_id"] ?>"
+                                                 <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?formation_category_id=<?php echo $row["formation_category_id"] ?>"
                                                     class="btn btn-outline-danger m-1">Delete</a>
                                             </td>
